+++ docs/src/examples/QasBoardGenerator/Basic.vue
@@ -1,25 +1,25 @@
 <template>
   <div class="q-py-md" style="height: 500px; background-color: var(--qas-background-color);">
     <qas-board-generator
-      v-model="boardColumns"
+      v-model:results="resultsColumns"
       column-id-key="id"
       column-url="url-example"
+      :headers="headers"
       height="400px"
-      :results="results"
     >
-      <template #header-column="{ context }">
+      <template #header-column="{ header }">
         <div class="flex justify-between text-subtitle1">
-          <div>{{ context.name }}</div>
+          <div>{{ header.name }}</div>
         </div>
       </template>
 
-      <template #column-item="{ context }">
+      <template #column-item="{ item }">
         <qas-box
           class="full-width q-mb-lg"
           style="height: 200px;"
         >
           <div class="q-mb-lg">
-            Card - {{ context.item }}
+            Card - {{ item.card }}
           </div>
         </qas-box>
       </template>
@@ -33,33 +33,33 @@ export default {
 
   data () {
     return {
-      boardColumns: {
+      resultsColumns: {
         /*
         * Somente de exemplo, quando na verdade o componente será responsável por bater
         * as APIs das colunas com base no headers fornecidos. É possível manipular as
         * colunas do lado de fora através deste model, como neste exemplo.
         */
         'col-1': [
-          { item: 'item 1 da coluna 1' },
-          { item: 'item 2 da coluna 1' },
-          { item: 'item 3 da coluna 1' }
+          { card: 'item 1 da coluna 1' },
+          { card: 'item 2 da coluna 1' },
+          { card: 'item 3 da coluna 1' }
         ],
         'col-2': [
-          { item: 'item 1 da coluna 2' },
-          { item: 'item 2 da coluna 2' },
-          { item: 'item 3 da coluna 2' }
+          { card: 'item 1 da coluna 2' },
+          { card: 'item 2 da coluna 2' },
+          { card: 'item 3 da coluna 2' }
         ],
         'col-3': [
-          { item: 'item 1 da coluna 3' },
-          { item: 'item 2 da coluna 3' },
-          { item: 'item 3 da coluna 3' }
+          { card: 'item 1 da coluna 3' },
+          { card: 'item 2 da coluna 3' },
+          { card: 'item 3 da coluna 3' }
         ]
       },
 
       /*
       * Headers na qual deverá fornecer para o componente.
       */
-      results: [
+      headers: [
         {
           id: 'col-1',
           name: 'Coluna 1'