+++ ui/src/components/board-generator/QasBoardGenerator.vue
@@ -1,13 +1,13 @@
 <template>
   <qas-grabbable class="qas-board-generator" use-scroll-bar>
     <div class="no-wrap q-col-gutter-sm q-px-xl row">
-      <div v-for="(header, index) in results" :key="index" class="q-mr-sm">
+      <div v-for="(header, index) in headers" :key="index" class="q-mr-sm">
         <qas-box class="q-mb-md">
-          <slot :context="header" name="header-column" />
+          <slot :fields="getFieldsByHeader(header)" :header="header" name="header-column" />
         </qas-box>
 
         <div ref="columnContainer" class="qas-board-generator__column secondary-scroll" :style="containerStyle">
-          <slot v-for="item in getItemsByHeader(header)" :context="item" name="column-item" />
+          <slot v-for="item in getItemsByHeader(header)" :fields="getFieldsByHeader(header)" :item="item" name="column-item" />
 
           <div class="full-width justify-center q-mb-md q-mt-sm row">
             <qas-btn v-if="hasSeeMore(header)" icon="sym_r_add" label="Ver mais" :use-label-on-small-screen="false" variant="tertiary" @click="fetchColumn(header)" />
@@ -23,7 +23,7 @@
 </template>
 
 <script setup>
-import { ref, watch, computed, onMounted, shallowRef, inject } from 'vue'
+import { ref, watch, computed, onMounted, markRaw, inject } from 'vue'
 import promiseHandler from '../../helpers/promise-handler'
 
 defineOptions({ name: 'QasBoardGenerator' })
@@ -31,13 +31,19 @@ defineOptions({ name: 'QasBoardGenerator' })
 const columnContainer = ref(null)
 const columnsPagination = ref({})
 const columnsLoading = ref({})
+const columnsFieldsModel = ref({})
 
 const axios = inject('axios')
 
 const props = defineProps({
-  results: {
+  headers: {
     type: Array,
     default: () => []
+  },
+
+  results: {
+    type: Object,
+    default: () => ({})
   },
 
   columnIdKey: {
@@ -55,11 +61,6 @@ const props = defineProps({
     required: true
   },
 
-  modelValue: {
-    type: Object,
-    default: () => ({})
-  },
-
   height: {
     type: String,
     default: ''
@@ -72,17 +73,30 @@ const props = defineProps({
 
   columnWidth: {
     type: String,
-    default: '288px'
+    default: '300px'
   },
 
-  useShallowRef: {
+  useMarkRaw: {
     type: Boolean,
     default: true
+  },
+
+  lazyLoadingFieldsKeys: {
+    type: Array,
+    default: () => []
   }
 })
 
+const emit = defineEmits([
+  'update:results',
+  'fetch-column-success',
+  'fetch-column-error',
+  'fetch-columns-success',
+  'fetch-columns-error'
+])
+
 watch(
-  () => props.results,
+  () => props.headers,
   () => {
     reset()
     setColumnHeightContainer()
@@ -98,21 +112,19 @@ onMounted(() => {
   fetchColumns()
 })
 
-const emit = defineEmits(['update:modelValue'])
-
 defineExpose({ fetchColumns, fetchColumn, reset })
 
-const columnsModel = computed({
+const columnsResultsModel = computed({
   get () {
-    return props.modelValue
+    return props.results
   },
 
   set (newValues) {
-    emit('update:modelValue', newValues)
+    emit('update:results', newValues)
   }
 })
 
-const hasColumnsLength = computed(() => Object.keys(columnsModel.value).length)
+const hasColumnsLength = computed(() => Object.keys(columnsResultsModel.value).length)
 
 const containerStyle = computed(() => `width: ${props.columnWidth};`)
 
@@ -134,7 +146,17 @@ function setColumnHeightContainer () {
 * Bater API pra cada header
 */
 async function fetchColumns () {
-  props.results.forEach(header => fetchColumn(header))
+  const promises = props.headers.map(header => fetchColumn(header))
+
+  const { error } = await promiseHandler(promises, { useLoading: false })
+
+  if (error) {
+    emit('fetch-columns-error', error)
+
+    return
+  }
+
+  emit('fetch-columns-success')
 }
 
 /*
@@ -161,32 +183,82 @@ async function fetchColumn (header) {
     }
   )
 
-  if (error) return
+  if (error) {
+    emit('fetch-column-error', error)
+
+    throw error
+  }
 
   /*
-  * exemplo de como columnsModel irá ficar:
+  * exemplo de como columnsResultsModel irá ficar:
   *
   * {
   *  '2024-02-15': [...],
   *  '2024-02-16': [...]
   * }
   *
-  * onde cada item do objeto é uma coluna no board. O mesmo vale para "columnsLoading" e "columnPagination", organizando
-  * os loadings e o controle de paginação por chave identificadora do header.
+  * onde cada item do objeto é uma coluna no board. O mesmo vale para "columnsFieldsModel", "columnsLoading" e
+  * "columnPagination", organizando os fields, loadings e o controle de paginação por chave identificadora do header.
   */
   const newColumnValues = [
-    ...columnsModel.value[headerKey] || [],
+    ...columnsResultsModel.value[headerKey] || [],
     ...response.data?.results || []
   ]
 
-  columnsModel.value[headerKey] = props.useShallowRef ? shallowRef(newColumnValues) : newColumnValues
+  columnsResultsModel.value[headerKey] = props.useMarkRaw ? markRaw(newColumnValues) : newColumnValues
 
-  columnsPagination.value[headerKey].offset = columnsModel.value[headerKey].length
+  /*
+  * Pode acontecer das options nos fields da segunda página serem diferentes da primeira página,
+  * portanto deve ocorrer o merge.
+  */
+  if (response.data?.fields) {
+    columnsFieldsModel.value[headerKey] = markRaw(
+      getMergedFields(columnsFieldsModel.value[headerKey], response.data?.fields)
+    )
+  }
+
+  columnsPagination.value[headerKey].offset = columnsResultsModel.value[headerKey].length
   columnsPagination.value[headerKey].count = response.data?.count
+
+  emit('fetch-column-success', { response, header })
+}
+
+/*
+* Mergeia os options antigos com os novos de cada field.
+*/
+function getMergedFields (oldFields, newFields) {
+  // Primeira vez batendo a API, retorna os novos fields.
+  if (!oldFields || !props.lazyLoadingFieldsKeys.length) return newFields
+
+  // Caso bata a API e por algum motivo não venha fields, mantenha o antigos.
+  if (oldFields && !newFields) return oldFields
+
+  const mergedFields = { ...oldFields }
+
+  props.lazyLoadingFieldsKeys.forEach(fieldKey => {
+    mergedFields[fieldKey].options = getNonDuplicatedOptions(oldFields[fieldKey].options, newFields[fieldKey].options)
+  })
+
+  return mergedFields
+}
+
+/*
+* Tratamento para fazer o merge e evitar options duplicados.
+*/
+function getNonDuplicatedOptions (oldOptions, newOptions) {
+  const options = [...oldOptions]
+
+  newOptions.forEach(item => {
+    const hasOption = options.find(option => option.value === item.value)
+
+    if (!hasOption) options.push(item)
+  })
+
+  return options
 }
 
 function getItemsByHeader (header) {
-  return hasColumnsLength.value ? columnsModel.value[getKeyByHeader(header)] : []
+  return hasColumnsLength.value ? columnsResultsModel.value[getKeyByHeader(header)] : []
 }
 
 /**
@@ -213,7 +285,7 @@ function setColumnsPagination () {
   columnsPagination.value = {}
   columnsLoading.value = {}
 
-  props.results.forEach(header => {
+  props.headers.forEach(header => {
     const headerKey = getKeyByHeader(header)
 
     columnsPagination.value[headerKey] = { limit: props.limitPerColumn, offset: 0 }
@@ -231,15 +303,21 @@ function hasEmptyResultText (header) {
 */
 function hasSeeMore (header) {
   const headerKey = getKeyByHeader(header)
-  const hasMorePagination = columnsModel.value[headerKey]?.length < columnsPagination.value[headerKey]?.count
+  const hasMorePagination = columnsResultsModel.value[headerKey]?.length < columnsPagination.value[headerKey]?.count
 
   return hasMorePagination && !columnsLoading.value[headerKey]
 }
 
 function reset () {
-  columnsModel.value = {}
+  columnsResultsModel.value = {}
   columnsPagination.value = {}
   columnsLoading.value = {}
+}
+
+function getFieldsByHeader (header) {
+  const headerKey = getKeyByHeader(header)
+
+  return columnsFieldsModel.value[headerKey] || {}
 }
 </script>